Broke browse dialog module into two
